Move BrowserClipboardService to /platform/ so it can be reused in the standalone editor
